added airport database functionality
new function "get_airport_stats()" added. The function takes in the college nickname as a parameter and returns a nested list of college airport stats. These are the indexes:
0: College full name
1: State of College
2: Name of Closest airport
3: Airport Code
4: DIstance of Airport from college

NOTE: Returned list from airport is a nested list hence while unpacking the list you will require to run two loops as a direct for loop on the returned list will just give an output of the nested list within the college.

The format of the list is given below:
[['Netaji Subhas University of Technology Delhi', 'Delhi', 'Indira Gandhi International Airport', 'DEL', '12.8']]